Fix minor warnings.
git-svn-id: https://svn.stepmania.com/svn/branches/4.0@28577 822b03c2-8f45-0410-a444-95660a741a6f

+++ stepmania/src/FileDownload.cpp
@@ -204,7 +204,7 @@ void FileTransfer::StartTransfer( TransferType type, const RString &sURL, const 
 	}
 	*/
 
-	vsHeaders.push_back( "Content-Length: " + ssprintf("%d",sRequestPayload.size()) );
+	vsHeaders.push_back( "Content-Length: " + ssprintf("%d",(int)sRequestPayload.size()) );
 
 	RString sHeader;
 	FOREACH_CONST( RString, vsHeaders, h )
